添加自定义转换的测试
